publish to Adafruit IO
- factor MQTT publishing into generalized Notifier class
- create Adafruit IO notifier
- create AWS IoT notifier

Adafruit IO is really sweet. The dashboard is a breeze to setup and the live
updating works great on laptop and phone.

Design note/todo:

thingpin uses the thingamon package to speak MQTT to AWS. To speak MQTT to
Adafruit IO thingpin uses the adafruit-io package. These two are very similar
and both use the same MQTT library underneath. The main difference how they
name topics.

I want to see if I can generalize thingpin a bit so that it can have
configurable topic patterns and then use it for both AWS IoT and Adafruit IO.